Refactor V3 pay modal
