+++ src/components/v2v3/V2V3Project/V2V3PayButton/V2V3ConfirmPayModal/hooks/NftRewardTiersToMint.tsx
@@ -1,0 +1,22 @@
+import { PayProjectFormContext } from 'components/Project/PayProjectForm/payProjectFormContext'
+import { NftRewardsContext } from 'contexts/nftRewardsContext'
+import { useContext } from 'react'
+import { rewardTiersFromIds } from 'utils/nftRewards'
+
+export function useNftRewardTiersToMint() {
+  const { form: payProjectForm } = useContext(PayProjectFormContext)
+  const {
+    nftRewards: { rewardTiers },
+  } = useContext(NftRewardsContext)
+
+  const nftTierIdsToMint = [
+    ...(payProjectForm?.payMetadata?.tierIdsToMint ?? []),
+  ].sort()
+
+  if (!rewardTiers || !nftTierIdsToMint) return
+
+  return rewardTiersFromIds({
+    tierIds: nftTierIdsToMint,
+    rewardTiers,
+  })
+}
